requirements_production.txt
# Production Requirements for Logistics AI Control System

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# AI/LLM Integration
openai==1.3.7
tenacity==8.2.3
structlog==23.2.0

# Mathematical Computing
sympy==1.12
numpy==1.24.4
scipy==1.11.4

# Web Framework
websockets==12.0
python-multipart==0.0.6

# Async Support
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# Data Processing
pandas==2.1.4
geopy==2.4.1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Optional: Background Tasks
celery==5.3.4
redis==5.0.1

# Optional: Caching
python-redis==5.0.1

# Optional: External APIs
requests==2.31.0
httpx==0.25.2

# Environment Management
python-dotenv==1.0.0